+++ src/FasTnT.Formatters.Xml/Formatters/XmlResponseFormatter.cs
@@ -1,9 +1,6 @@
 ﻿using FasTnT.Commands.Responses;
-using FasTnT.Parsers.Xml.Formatters.Implementation;
 using FasTnT.Parsers.Xml.Utils;
 using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Xml.Linq;
 
@@ -14,29 +11,7 @@ namespace FasTnT.Parsers.Xml.Formatters
         public override XDocument FormatInternal(PollResponse response)
         {
             var formatted = CreateResponse("EPCISQueryDocument");
-            var resultName = "EventList";
-            var resultList = default(IEnumerable<XElement>);
-
-            if (response.EventList.Any())
-            {
-                resultName = "EventList";
-                resultList = XmlEntityFormatter.FormatEvents(response.EventList);
-            }
-            else if (response.MasterdataList.Any())
-            {
-                resultName = "VocabularyList";
-                resultList = XmlEntityFormatter.FormatMasterData(response.MasterdataList);
-            }
-
-            formatted.Root.Add(
-                new XElement("EPCISBody",
-                    new XElement(XName.Get("QueryResults", EpcisNamespaces.Query),
-                        new XElement("queryName", response.QueryName),
-                        !string.IsNullOrEmpty(response.SubscriptionId) ? new XElement("subscriptionID", response.SubscriptionId) : null,
-                        new XElement("resultsBody", new XElement(resultName, resultList))
-                    )
-                )
-            );
+            formatted.Root.Add(new XElement("EPCISBody", FormatPollResponse(response)));
 
             return formatted;
         }
@@ -51,26 +26,23 @@ namespace FasTnT.Parsers.Xml.Formatters
 
         public override XDocument FormatInternal(GetSubscriptionIdsResponse response)
         {
-            var formatted = CreateResponse("EPCISQueryDocument");
-            formatted.Root.Add(new XElement("EPCISBody", new XElement(XName.Get("GetSubscriptionIDsResult", EpcisNamespaces.Query), response.SubscriptionIds.Select(x => new XElement("string", x)))));
-
-            return formatted;
+            return CreateQueryResponse(XName.Get("GetSubscriptionIDsResult", EpcisNamespaces.Query), response.SubscriptionIds.Select(x => new XElement("string", x)));
         }
 
         public override XDocument FormatInternal(GetStandardVersionResponse response)
         {
-            var formatted = CreateResponse("EPCISQueryDocument");
-            formatted.Root.Add(new XElement("EPCISBody", new XElement(XName.Get("GetStandardVersionResult", EpcisNamespaces.Query), response.Version)));
-
-            return formatted;
+            return CreateQueryResponse(XName.Get("GetStandardVersionResult", EpcisNamespaces.Query), response.Version);
         }
 
         public override XDocument FormatInternal(GetQueryNamesResponse response)
         {
+            return CreateQueryResponse(XName.Get("GetQueryNamesResult", EpcisNamespaces.Query), response.QueryNames.Select(x => new XElement("string", x)));
+        }
+
+        private XDocument CreateQueryResponse(XName element, object content)
+        {
             var formatted = CreateResponse("EPCISQueryDocument");
-            formatted.Root.Add(
-                new XElement("EPCISBody", new XElement(XName.Get("GetQueryNamesResult", EpcisNamespaces.Query), response.QueryNames.Select(x => new XElement("string", x))))
-            );
+            formatted.Root.Add(new XElement("EPCISBody", new XElement(element, content)));
 
             return formatted;
         }
